requirements.txt
boto3==1.34.11
python-dotenv==1.0.0
streamlit==1.29.0
pandas==2.1.4